refactor: view input/output 하나로 합치기
- inputView에서 nextInt뒤에 nextLine이 동작하지 않는 에러 수정을 위해 nextInt를 Integer.parseInt(scanner.nextLine())으로 수정
- money의 수익률을 구하는 과정을 Money에 getYield메서드를 통해 값을 얻어오는 것으로 변경
